django-pfiProject-docker/requirements.txt
Django==2.2.3
djangorestframework==3.8
djangorestframework-gis==0.14.0
django-filter==2.2
djangorestframework-filters>=0.11.1.dev0
djangorestframework-queryfields
#django-cors-headers
django-map-widgets
markdown
packaging
six
psycopg2-binary==2.7.5
python-dotenv==0.9.1
pytz==2018.5
uwsgi==2.0.18
django-allauth==0.38.0
kafka-python==1.4.4
